# cheeseAPI/swagger.yaml
openapi: 3.0.0
info:
  title: Cheeses API
  version: 1.0.0
  description: API documentation for a simple cheeses API
servers:
  - url: http://localhost:3000
paths:
  /cheeses:
    get:
      summary: Get all cheeses
      tags:
        - Cheeses
      responses:
        '200':
          description: A list of cheeses
    post:
      summary: Create a new cheese
      tags:
        - Cheeses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - picture
                - color
                - price
              properties:
                name:
                  type: string
                  example: "Cheddar"
                picture:
                  type: string
                  example: "http://example.com/cheddar.jpg"
                color:
                  type: string
                  example: "Yellow"
                price:
                  type: number
                  format: float
                  example: 5.99
      responses:
        '201':
          description: Cheese created
  /cheeses/{id}:
    get:
      summary: Get a cheese by ID
      tags:
        - Cheeses
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the cheese
          schema:
            type: integer
      responses:
        '200':
          description: The cheese
        '404':
          description: Cheese not found
    put:
      summary: Update a cheese by ID
      tags:
        - Cheeses
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the cheese
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - picture
                - color
                - price
              properties:
                name:
                  type: string
                  example: "Aged Cheddar"
                picture:
                  type: string
                  example: "http://example.com/aged-cheddar.jpg"
                color:
                  type: string
                  example: "Dark Yellow"
                price:
                  type: number
                  format: float
                  example: 6.99
      responses:
        '200':
          description: Cheese updated
        '404':
          description: Cheese not found
    delete:
      summary: Delete a cheese by ID
      tags:
        - Cheeses
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the cheese
          schema:
            type: integer
      responses:
        '204':
          description: Cheese deleted
        '404':
          description: Cheese not found